SOA/DS2/requirements.txt
faker==38.2.0
numpy==2.3.5
matplotlib==3.10.7
requests==2.32.3
python-dotenv==1.2.1
djangorestframework-simplejwt==5.3.1
scikit-learn==1.6.1
Django==5.1.4
djangorestframework==3.15.2